Separate Config/Constants from Business Logic in authController
