+++ backend/config/authConfig.js
@@ -1,0 +1,23 @@
+const ACCESS_TOKEN_EXPIRY = "15m";
+const REFRESH_TOKEN_EXPIRY = "30d";
+const REFRESH_TOKEN_MAX_AGE_MS = 30 * 24 * 60 * 60 * 1000;
+const REFRESH_TOKEN_SALT_ROUNDS = 10;
+
+const getRefreshCookieOptions = () => ({
+    httpOnly: true,
+    secure: process.env.NODE_ENV === "production",
+    sameSite: process.env.NODE_ENV === "production" ? "None" : "Lax",
+    maxAge: REFRESH_TOKEN_MAX_AGE_MS,
+    path: "/api/auth",
+});
+
+const EMAIL_REGEX = /^[^\s@]+@[^\s@]+\.[A-Za-z]{2,}$/;
+
+module.exports = {
+    ACCESS_TOKEN_EXPIRY,
+    REFRESH_TOKEN_EXPIRY,
+    REFRESH_TOKEN_MAX_AGE_MS,
+    REFRESH_TOKEN_SALT_ROUNDS,
+    getRefreshCookieOptions,
+    EMAIL_REGEX,
+};
