+++ backend/controllers/authController.js
@@ -4,18 +4,13 @@ const jwt = require("jsonwebtoken");
 const crypto = require("crypto");
 const { sendVerificationEmail } = require("../utils/sendEmail");
 const { validatePassword } = require('../utils/passwordPolicy');
-
-const ACCESS_TOKEN_EXPIRY = "15m";
-const REFRESH_TOKEN_EXPIRY = "30d";
-const REFRESH_TOKEN_MAX_AGE_MS = 30 * 24 * 60 * 60 * 1000;
-const REFRESH_TOKEN_SALT_ROUNDS = 10;
-const getRefreshCookieOptions = () => ({
-    httpOnly: true,
-    secure: process.env.NODE_ENV === "production",
-    sameSite: process.env.NODE_ENV === "production" ? "None" : "Lax",
-    maxAge: REFRESH_TOKEN_MAX_AGE_MS,
-    path: "/api/auth",
-});
+const { 
+    ACCESS_TOKEN_EXPIRY, 
+    REFRESH_TOKEN_EXPIRY, 
+    REFRESH_TOKEN_SALT_ROUNDS, 
+    getRefreshCookieOptions, 
+    EMAIL_REGEX 
+} = require("../config/authConfig");
 
 /**
  * Generate an access token for the authenticated user.
@@ -45,9 +40,7 @@ const registerUser = async (req, res) => {
     try {
         const { name, email, password, profileImageUrl } = req.body;
         
-        const emailRegex = /^[^\s@]+@[^\s@]+\.[A-Za-z]{2,}$/;
-
-        if (!emailRegex.test(email)) {
+        if (!EMAIL_REGEX.test(email)) {
            return res.status(400).json({
            success: false,
            message: "Please enter a valid email address.",
